tentando fazer com que quando existir uma imagem ja com mesmo nome elaseja deletada do diretorio e seja dado update na tabela atravez do seu id;

+++ admin/index.php
@@ -6,13 +6,16 @@ $objFunc = new Funcionario();
 //VALIDANDO USUARIO
 session_start();
 
+$objFunc->trasImagem();
+
+$foto = $_SESSION['foto'];
+$idfoto = $_SESSION['idfoto'];
 
 
 if($_SESSION["logado"] == "sim"){
         
 	$objFunc->funcionarioLogado($_SESSION['func']);
         $id = $_SESSION['func'];
-        $foto = $_SESSION['fotoperfil'];
         
 }else{
 	header("location: http://localhost/crud_phpoo/"); 
@@ -85,32 +88,68 @@ if(isset($_POST['enviaimg'])){
 					$nome_final = $_FILES['imagem']['name'];
 				}
 				//Verificar se é possivel mover o arquivo para a pasta escolhida
-				if(move_uploaded_file($_FILES['imagem']['tmp_name'], $_UP['pasta']. $nome_final)){
-					//Upload efetuado com sucesso, exibe a mensagem
-                                        //
+                                
+                                if(file_exists('../upload/perfil/'.$foto)){                                    
+                                    unlink('../upload/perfil/'.$foto);
                                     
-                                        $objFunc->updateFoto($nome_final);
+                                    if(move_uploaded_file($_FILES['imagem']['tmp_name'], $_UP['pasta']. $nome_final)){
+					//Upload efetuado com sucesso, exibe a mensagem
+                                        //                                    
+                                        $objFunc->iserirFoto($nome_final);
+                                        
 //					$query = mysqli_query($conn, "INSERT INTO usuarios (
 //					nome_imagem) VALUES('$nome_final')");
-//					echo "
-//						<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://localhost/Aula/upload_imagem.php'>
-//						<script type=\"text/javascript\">
-//							alert(\"Imagem cadastrada com Sucesso.\");
-//						</script>
-//					";	
-//				}else{
-//					//Upload não efetuado com sucesso, exibe a mensagem
-//					echo "
-//						<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://localhost/crud_phpoo/admin/index.php?error=true'>
-//						<script type=\"text/javascript\">
-//							alert(\"Imagem não foi cadastrada com Sucesso.\");
-//						</script>
-//					";
+					echo "
+						<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://localhost/crud_phpoo/admin/index.php'>
+						<script type=\"text/javascript\">
+							alert(\"Imagem cadastrada com Sucesso.\");
+						</script>
+					";
+                                        
+				}else{
+					//Upload não efetuado com sucesso, exibe a mensagem
+					echo "
+						<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://localhost/crud_phpoo/admin/index.php?error=true'>
+						<script type=\"text/javascript\">
+							alert(\"Imagem não foi cadastrada com Sucesso.\");
+						</script>
+					";
+				}
+                                    
+                                    
+                                    
+                                    
+                                }else{
+                                    
+                                
+				if(move_uploaded_file($_FILES['imagem']['tmp_name'], $_UP['pasta']. $nome_final)){
+					//Upload efetuado com sucesso, exibe a mensagem
+                                        //                                    
+                                        $objFunc->updateFoto($nome_final, $idfoto);
+                                        
+//					$query = mysqli_query($conn, "INSERT INTO usuarios (
+//					nome_imagem) VALUES('$nome_final')");
+					echo "
+						<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://localhost/crud_phpoo/admin/index.php'>
+						<script type=\"text/javascript\">
+							alert(\"Imagem cadastrada com Sucesso.\");
+						</script>
+					";
+                                        
+				}else{
+					//Upload não efetuado com sucesso, exibe a mensagem
+					echo "
+						<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://localhost/crud_phpoo/admin/index.php?error=true'>
+						<script type=\"text/javascript\">
+							alert(\"Imagem não foi cadastrada com Sucesso.\");
+						</script>
+					";
 				}
 			}
     
-	
+                                }
 }
+                        
 
 if(isset($_GET['sair']) == "sim"){
 	$objFunc->sairFuncionario();
@@ -163,11 +202,11 @@ if(isset($_GET['sair']) == "sim"){
     </ul>
   </div>
 </nav>
-    <?php echo $foto; ?>
+    
     <?php
         
 
-    if (!empty($foto) == ''){
+    if (!empty($foto[0]) == ''){
        
     echo '<div class="container">
         <a id="FTperfil" data-toggle="modal" data-target="#myModal"><img width="120px" height="120px" src="../upload/perfil/default.png" /></a><br>
